+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to the "arb-translator" extension will be documented in this file.
 
+## [1.3.5] - 24.01.23
+### Added
+- Add `Delete key` command.
+
+### Updated
+- Update history when key changings.
+
 ## [1.3.4] - 24.01.23
 ### Added
 - Add `Change key` command.